Mior update to morph import functions
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@1419 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/cell/converters/SWCMorphReader.java
@@ -153,7 +153,7 @@ public class SWCMorphReader extends FormatImporter
                         
                         try
                         {
-                            sectionType = Integer.parseInt(items[1]);
+                            sectionType = (int)Float.parseFloat(items[1]);
                         }
                         catch (NumberFormatException e)
                         {
